Refactor CreateDatabaseDiscoveryRuleStatementUpdaterTest
